+++ include/TombRaider.h
@@ -11,6 +11,12 @@
 
 #include <cstdint>
 
+#ifdef WIN32
+typedef uint8_t u_int8_t;
+typedef uint16_t u_int16_t;
+typedef uint32_t u_int32_t;
+#endif
+
 #include "TombRaiderData.h"
 
 /*!