+++ src/utils/time.cpp
@@ -5,9 +5,11 @@
  * \author xythobuz
  */
 
-#include <sys/time.h>
-
 #include "utils/time.h"
+
+#if defined(unix) || defined(__APPLE__) || defined (__linux__)
+
+#include <sys/time.h>
 
 struct timeval system_timer_start;
 struct timeval system_timer_stop;
@@ -22,4 +24,27 @@ unsigned long systemTimerGet() {
 void systemTimerReset() {
     gettimeofday(&system_timer_start, &system_timer_zone);
 }
+
+#elif defined(WIN32)
+
+#include <Windows.h>
+
+DWORD system_timer_start = 0;
+
+unsigned long systemTimerGet() {
+    return GetTickCount() - system_timer_start;
+}
+
+void systemTimerReset() {
+    system_timer_start = GetTickCount();
+}
+
+#else
+#warn "No support for timer on this platform!"
+
+unsigned long systemTimerGet() { return 0; }
+
+void systemTimerReset() { }
+
+#endif